refactor: extract execution graph logic into dedicated module and hook
