+++ packages/plugins/plugin-assistant/src/hooks/index.ts
@@ -13,3 +13,4 @@ export * from './useOnline';
 export * from './usePresets';
 export * from './useReferencesProvider';
 export * from './useResolveRef';
+export * from './useExecutionGraph';
